refactor: Move SetupChainParamsBaseOptions to util/args
To eliminate circular chainparamsbase dependencies.
