+++ src/chainparams.h
@@ -17,6 +17,8 @@
 #include <string>
 #include <vector>
 
+class ArgsManager;
+
 typedef std::map<int, uint256> MapCheckpoints;
 
 struct CCheckpointData {
