+++ src/chainparamsbase.h
@@ -8,7 +8,7 @@
 #include <memory>
 #include <string>
 
-class ArgsManager;
+// @see SetupChainParamsBaseOptions for ArgsManager interation
 
 /**
  * CBaseChainParams defines the base parameters (shared between bitcoin-cli and bitcoind)
@@ -45,11 +45,6 @@ private:
  * @throws a std::runtime_error if the chain is not supported.
  */
 std::unique_ptr<CBaseChainParams> CreateBaseChainParams(const std::string& chain);
-
-/**
- *Set the arguments for chainparams
- */
-void SetupChainParamsBaseOptions(ArgsManager& argsman);
 
 /**
  * Return the currently selected parameters. This won't change after app
