+++ test/lint/lint-circular-dependencies.sh
@@ -9,8 +9,6 @@
 export LC_ALL=C
 
 EXPECTED_CIRCULAR_DEPENDENCIES=(
-    "chainparamsbase -> util/args -> chainparamsbase"
-    "chainparamsbase -> util/args -> util/system -> chainparamsbase"
     "node/blockstorage -> validation -> node/blockstorage"
     "index/blockfilterindex -> node/blockstorage -> validation -> index/blockfilterindex"
     "index/base -> validation -> index/blockfilterindex -> index/base"
